backend/requirements.txt
# Django and REST Framework
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.6.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.10.1

# API Documentation and Utilities
drf-spectacular==0.27.2
django-filter==24.3
python-decouple==3.8

# Database (for production)
# psycopg2-binary==2.9.9

# Development dependencies
Pillow==10.4.0
